Implemented validating user permissions when deleting messages ( #23 )

+++ LiteChat/LiteChat/src/app/components/question/question.component.ts
@@ -1,6 +1,7 @@
 import { Component, Input } from '@angular/core';
 import { QuestionService } from 'src/app/core/services/question/question.service';
 import { Message } from '../../core/models/message';
+import { LocalStorage } from 'src/environments/LocalStorage';
 @Component({
   selector: 'app-question',
   templateUrl: './question.component.html',
@@ -8,13 +9,23 @@ import { Message } from '../../core/models/message';
 })
 export class QuestionComponent {
   @Input() message!: Message;
-  isAdmin: boolean = false;
-  constructor(private questionService: QuestionService) {}
+  isAdmin: boolean = this.localStorage.data.IsAdmin;
+  constructor(
+    private questionService: QuestionService,
+    private localStorage: LocalStorage
+  ) {}
 
-  //VALIDATE USER PERMISSIONS
   deleteMessage(message: Message) {
-    if (confirm('Are you sure you want to delete this question?'))
-      this.questionService.deleteMessage(message);
+    if (this.isAdmin) {
+      this.questionService
+        .isUserAdmin(this.localStorage.data.ApplicationUserId)
+        .subscribe((isAdmin) => {
+          if (isAdmin) {
+            if (confirm('Are you sure you want to delete this question?'))
+              this.questionService.deleteMessage(message);
+          }
+        });
+    }
   }
 
   likeMessage(message: Message) {
